requirements.txt
cachetools==5.3.0
charset-normalizer==3.0.1
click==8.1.3
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.77.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
imageio==2.25.0
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
moviepy==1.0.3
numpy==1.24.2
oauth2client==4.1.3
opencv-python-headless==4.7.0.68
Pillow==9.4.0
proglog==0.1.10
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
pyparsing==3.0.9
pyTelegramBotAPI==4.10.0
python-dotenv==0.21.1
pytube==12.1.2
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.2
